fix(pricing): make catalog reconciliation race-safe
Move catalog reconciliation out of request-time pricing lookup and make store updates reject stale concurrent plans.
