+++ crates/gateway-service/src/pricing_catalog/tests/mod.rs
@@ -1,0 +1,4 @@
+mod reconciliation;
+mod refresh;
+mod resolution;
+mod support;
